Fix Enumerator returned from KDTreeNodeCollection
Currently, the enumerator uses `position.Length` to determine how many items to iterate over.  This results in the enumerator returning objects that don't exist if the collection isn't full (as can happen when using `Nearest(location, radius, maxCount)`).

This fix changes the enumerator to use the already-existing `count` variable, which ought to prevent it from returning non-existent collection elements.