Update RandomForest.cs
This fixes a problem with serializing a RandomForest object. ParallelOptions cannot be serialized and thus needs to be excluded from the serialization process. In order to assure functionality after deserialization an internal method (OnDeserializationMethod) reinitializes the ParallelOptions parameter. I tested serialization and deserialization and it seems to be working fine now.