Update SequentialMinimalOptimizationRegression.cs
This updates allows users to set a value of the ε parameter really. It has been tested using the Regression (SVMs) program from the samples. Without this changes it was impossible because the private variable epsilon of the BaseSequentialMinimalOptimizationRegression class used the fixed value 1e-3 and not changed whit changing the Epsilon property of the BaseSupportVectorRegression parent class.