GH-752: Removing redundant array allocation from rowVector * matrix calculation.
It is common to cache a column in a row-major language like C# for GEMM operations but it is not needed for vector-matrix operations as the cached column is never reused. It is simply assigned and then used. In terms of performance, this is more than twice as fast for N<=128 but the speed improvement gets smaller for large N as the relative cost of allocation decreases.