Remove got_packet check
The got_packet checks fails consistently at start when encoding with the H264 encoder. When this happens, the same frame is encoded several times avcodec_encode_video2. This will result in duplicated frames at start of the output video, as well as the error message "non monotonic PTS" for the first couple of frames.
Some more research might be required in how got_packet should be handled, however as I am experiencing myself, removing the check works better than leaving it (as the audio codec doesn't have this problem, audio and video will be unaligned)